# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Micro Wallet Documentation",
        "description": "Software for manage wallets and transictions by users",
        "contact": {
            "email": "menezes.adolfo@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api/v1",
            "description": "API Documentation for Micro Wallet"
        }
    ],
    "paths": {
...
        "/wallets": {
            "get": {
                "tags": [
                    "Wallets"
                ],
                "summary": "Get list of wallets",
                "description": "Returns list of wallets",
                "operationId": "getAllWallets",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "list of wallets",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number"
                                                    },
                                                    "user_id": {
                                                        "type": "number"
                                                    },
                                                    "balance": {
                                                        "type": "number"
                                                    },
                                                    "status": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Wallets"
                ],
                "summary": "Create wallet",
                "description": "Return create new wallet",
                "operationId": "storeWallet",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "number"
                                    },
                                    "balance": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "data of new wallet",
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "number"
                                                },
                                                "user_id": {
                                                    "type": "number"
                                                },
                                                "balance": {
                                                    "type": "number"
                                                },
                                                "status": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Server Erro",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/wallets/{id}": {
            "get": {
                "tags": [
                    "Wallets"
                ],
                "summary": "List wallet by id",
                "description": "Returns data of wallet",
                "operationId": "getWalletsById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "wallet id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "data of wallet",
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "number"
                                                },
                                                "user_id": {
                                                    "type": "number"
                                                },
                                                "balance": {
                                                    "type": "number"
                                                },
                                                "status": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Server Erro",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Wallets"
                ],
                "summary": "Update wallet",
                "description": "Return update wallet",
                "operationId": "updateWallet",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "wallet id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "number"
                                    },
                                    "balance": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "data of updated wallet",
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "number"
                                                },
                                                "user_id": {
                                                    "type": "number"
                                                },
                                                "balance": {
                                                    "type": "number"
                                                },
                                                "status": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Server Erro",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Wallets"
                ],
                "summary": "Delete wallet",
                "description": "Return delete wallet",
                "operationId": "deleteWallet",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "wallet id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Server Erro",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        }
    },
...
    "security": [
        []
    ]
}